Secondary weapon

+++ Assets/Scripts/Player/Player.cs
@@ -1,14 +1,16 @@
+using System;
 using System.Collections;
 using System.Collections.Generic;
 using UnityEngine;
+using UnityEngine.Events;
 
 namespace Player
-{
+{   
     public class Player : MonoBehaviour
     {
-        private int _health;
+        private float _health;
 
-        public void TakeDamage(int damage)
+        public void TakeDamage(float damage)
         {
             _health -= damage;
         }
